+++ ocdsextensionregistry/exceptions.py
@@ -10,5 +10,9 @@ class MissingExtensionMetadata(OCDSExtensionRegistryError):
     """Raised if a method call requires extensions metadata, with which the extension registry was not initialized"""
 
 
+class NotAvailableInBulk(OCDSExtensionRegistryError):
+    """Raised if files are required to be available in bulk, but are not"""
+
+
 class CommandError(OCDSExtensionRegistryError):
     """Errors from within this package's CLI"""